Refactor RedirectUrl extension to use controller structure
